🚚 add pure to function/file names
