+++ packages/atom.io/src/internal/families/index.ts
@@ -1,9 +1,9 @@
 export * from "./create-atom-family"
-export * from "./create-readonly-selector-family"
-export * from "./create-readonly-selector-family"
+export * from "./create-readonly-pure-selector-family"
+export * from "./create-readonly-pure-selector-family"
 export * from "./create-regular-atom-family"
 export * from "./create-selector-family"
-export * from "./create-writable-selector-family"
+export * from "./create-writable-pure-selector-family"
 export * from "./dispose-from-store"
 export * from "./find-in-store"
 export * from "./init-family-member"
