+++ packages/atom.io/src/internal/selector/index.ts
@@ -1,6 +1,6 @@
-export * from "./create-readonly-selector"
+export * from "./create-readonly-pure-selector"
 export * from "./create-standalone-selector"
-export * from "./create-writable-selector"
+export * from "./create-writable-pure-selector"
 export * from "./dispose-selector"
 export * from "./get-selector-dependency-keys"
 export * from "./register-selector"
